fix pages
