+++ lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:asia_project/views/reports_admin_view.dart';
+import 'package:asia_project/views/reports_coders_views.dart';
 import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'firebase_options.dart'; 
@@ -21,7 +22,7 @@ class MyApp extends StatelessWidget {
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),
-      home: const ReportsAdmin(), 
+      home: const ReportsCoders(), 
     );
   }
 }
