+++ lib/views/reports_admin_view.dart
@@ -1,5 +1,5 @@
 import 'package:asia_project/widgets/reports_bi_widgets/bar_chart.dart';
-import 'package:asia_project/widgets/reports_bi_widgets/filter_admin.dart';
+import 'package:asia_project/widgets/reports_bi_widgets/filters.dart';
 import 'package:asia_project/widgets/reports_bi_widgets/coder_table.dart';
 import 'package:asia_project/widgets/reports_bi_widgets/header_admin_widget.dart';
 import 'package:asia_project/widgets/reports_bi_widgets/search_input_widget.dart';
@@ -59,13 +59,13 @@ class _ReportsAdminState extends State<ReportsAdmin> {
               padding: EdgeInsets.all(16.0),
               child: SearchInputWidget(),
             ),
-            const FilterAdmin(),
-            StudentTable(),
+            const Filters(),
             BarChartWidget(
               chartTitle: "Attendance Chart",
               ref: attendanceChartRefs,
               data: attendanceChartData,
             ),
+            StudentTable(),
           ],
         ),
       ),
