+++ lib/views/reports_coders_views.dart
@@ -1,18 +1,17 @@
 import 'package:asia_project/widgets/reports_bi_widgets/bar_chart.dart';
-import 'package:asia_project/widgets/reports_bi_widgets/filter_admin.dart';
+import 'package:asia_project/widgets/reports_bi_widgets/filters.dart';
 import 'package:asia_project/widgets/reports_bi_widgets/coder_table.dart';
-import 'package:asia_project/widgets/reports_bi_widgets/search_input_widget.dart';
 import 'package:asia_project/widgets/reports_bi_widgets/header_coder_widget.dart';
 import 'package:flutter/material.dart';
 
-class ReportsAdmin extends StatefulWidget {
-  const ReportsAdmin({super.key});
+class ReportsCoders extends StatefulWidget {
+  const ReportsCoders({super.key});
 
   @override
-  State<ReportsAdmin> createState() => _ReportsAdminState();
+  State<ReportsCoders> createState() => _ReportsCodersState();
 }
 
-class _ReportsAdminState extends State<ReportsAdmin> {
+class _ReportsCodersState extends State<ReportsCoders> {
   final Map<String, String> attendanceChartRefs = {
     'titleFirstValue': 'Justified Absence',
     'titleSecondValue': 'Unjustified Absence',
@@ -57,13 +56,13 @@ class _ReportsAdminState extends State<ReportsAdmin> {
             const Padding(
               padding: EdgeInsets.all(16.0),
             ),
-            const FilterAdmin(),
-            StudentTable(),
+            const Filters(),
             BarChartWidget(
               chartTitle: "Attendance Chart",
               ref: attendanceChartRefs,
               data: attendanceChartData,
             ),
+            StudentTable(),
           ],
         ),
       ),
