+++ lib/widgets/reports_bi_widgets/coder_card_widget.dart
@@ -32,6 +32,7 @@ class StudentCard extends StatelessWidget {
       margin: EdgeInsets.only(bottom: 16),
       elevation: 4,
       shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
+      color: Colors.white,
       child: Padding(
         padding: const EdgeInsets.all(16.0),
         child: Row(
